+++ xarray/core/dataarray.py
@@ -4418,21 +4418,21 @@ class DataArray(AbstractArray, DataWithCoords):
         )
         return ds[self.name]
 
-    def drop_duplicates(
+    def drop_duplicate_coords(
         self,
-        dims: Union[Hashable, Iterable[Hashable]] = None,
+        coords: Union[Hashable, Iterable[Hashable]] = ...,
         keep: Union[
             str,
             bool,
         ] = "first",
     ):
-        """Returns a new DataArray with duplicate dimension values removed.
+        """Returns a new DataArray with duplicate coordinate values removed.
 
         Parameters
         ----------
-        dims : dimension label or sequence of labels, optional
-            Only consider certain dimensions for identifying duplicates, by
-            default use all dimensions.
+        coords : coordinate label or sequence of labels, optional
+            Only consider certain coordinates for identifying duplicates, by
+            default use all dimensions (NOT coordinates).
         keep : {"first", "last", False}, default: "first"
             Determines which duplicates (if any) to keep.
             - ``"first"`` : Drop duplicates except for the first occurrence.
@@ -4443,7 +4443,7 @@ class DataArray(AbstractArray, DataWithCoords):
         -------
         DataArray
         """
-        ds = self._to_temp_dataset().drop_duplicates(dims=dims, keep=keep)
+        ds = self._to_temp_dataset().drop_duplicate_coords(coords=coords, keep=keep)
         return self._from_temp_dataset(ds)
 
     # this needs to be at the end, or mypy will confuse with `str`