+++ xarray/tests/test_dataset.py
@@ -6848,7 +6848,7 @@ def test_deepcopy_obj_array():
 
 
 @pytest.mark.parametrize("keep", ["first", "last", False])
-def test_drop_duplicates(keep):
+def drop_duplicate_coords_dim(keep):
     ds = xr.DataArray(
         [0, 5, 6, 7], dims="time", coords={"time": [0, 0, 1, 2]}, name="test"
     ).to_dataset()
@@ -6866,12 +6866,12 @@ def test_drop_duplicates(keep):
     expected = xr.DataArray(
         data, dims="time", coords={"time": time}, name="test"
     ).to_dataset()
-    result = ds.drop_duplicates("time", keep=keep)
+    result = ds.drop_duplicate_coords("time", keep=keep)
     assert_equal(expected, result)
 
 
 @pytest.mark.parametrize("keep", ["first", "last", False])
-def test_drop_duplicates_multi_dim(keep):
+def test_drop_duplicate_coords_dims(keep):
     base_data = np.stack([np.arange(0, 5) * i for i in np.arange(0, 5)])
     ds = xr.DataArray(
         base_data,
@@ -6896,5 +6896,66 @@ def test_drop_duplicates_multi_dim(keep):
     expected = xr.DataArray(
         data, dims=["lat", "lon"], coords={"lat": lat, "lon": lon}, name="test"
     ).to_dataset()
-    result = ds.drop_duplicates(["lat", "lon"], keep=keep)
+    result = ds.drop_duplicate_coords(["lat", "lon"], keep=keep)
     assert_equal(expected, result)
+
+
+@pytest.mark.parametrize("keep", ["first", "last", False])
+def test_drop_duplicate_coords(keep):
+    ds = xr.DataArray(
+        [["a", "b", "c"], ["d", "e", "f"]],
+        coords={"init": [0, 1], "tau": [1, 2, 3]},
+        dims=["init", "tau"],
+    ).to_dataset(name="test")
+    ds.coords["valid"] = (("init", "tau"), np.array([[8, 6, 6], [7, 7, 7]]))
+
+    if keep == "first":
+        data = [["a", "b"], ["d", np.nan]]
+        init = [0, 1]
+        tau = [1, 2]
+        valid = [8.0, 7.0, 6.0, np.nan]
+    elif keep == "last":
+        data = [["a", "c"], [np.nan, "f"]]
+        init = [0, 1]
+        tau = [1, 3]
+        valid = [8.0, 6.0, np.nan, 7]
+    else:
+        data = [["a"]]
+        init = [0]
+        tau = [1]
+        valid = [8]
+
+    result = ds.drop_duplicate_coords("valid", keep=keep)
+    expected = xr.DataArray(
+        data,
+        dims=["init", "tau"],
+        coords={"init": init, "tau": tau, "valid": valid},
+        name="test",
+    ).to_dataset()
+    assert_equal(expected, result)
+
+
+@pytest.mark.parametrize("keep", ["first", "last", False])
+def test_drop_duplicate_coords_duplicate_dims(keep):
+    ds = xr.DataArray(
+        [["a", "b", "c"], ["d", "e", "f"]],
+        coords={"init": [0, 0], "tau": [1, 2, 3]},  # duplicate inits
+        dims=["init", "tau"],
+    ).to_dataset(name="test")
+    ds.coords["valid"] = (("init", "tau"), np.array([[8, 6, 6], [7, 7, 7]]))
+
+    with pytest.raises(ValueError):
+        ds.drop_duplicate_coords("valid", keep=keep)
+
+
+@pytest.mark.parametrize("keep", ["first", "last", False])
+def test_drop_duplicate_coords_missing(keep):
+    ds = xr.DataArray(
+        [["a", "b", "c"], ["d", "e", "f"]],
+        coords={"init": [0, 0], "tau": [1, 2, 3]},  # duplicate inits
+        dims=["init", "tau"],
+    ).to_dataset(name="test")
+    ds.coords["valid"] = (("init", "tau"), np.array([[8, 6, 6], [7, 7, 7]]))
+
+    with pytest.raises(ValueError):
+        ds.drop_duplicate_coords("doesnt exist", keep=keep)